application/requirements.txt
psycopg2-binary==2.9.5
flask==2.2.2